+++ Package.swift
@@ -13,7 +13,7 @@ let package = Package(
     dependencies: [
         .package(url: "https://github.com/swiftlang/swift-docc-plugin.git", from: "1.4.0"),
         .package(url: "https://github.com/bare-swift/swift-bytes.git", from: "0.1.0"),
-        .package(url: "https://github.com/bare-swift/swift-deflate.git", from: "0.4.0")
+        .package(url: "https://github.com/bare-swift/swift-deflate.git", from: "0.5.0")
     ],
     targets: [
         .target(